+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [Bugfix: fix feign report call result error when using feign direct call](https://github.com/Tencent/spring-cloud-tencent/pull/621)
 - [fix pr 606: modify a part of changes requested by review.](https://github.com/Tencent/spring-cloud-tencent/pull/620)
 - [fix pr 613: modify a judgment logic](https://github.com/Tencent/spring-cloud-tencent/pull/618)
+- [Feature: support new label expression](https://github.com/Tencent/spring-cloud-tencent/pull/627)